% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nproc.R
\name{plot.nproc}
\alias{plot.nproc}
\title{Plot the nproc band(s).}
\usage{
\method{plot}{nproc}(x, ...)
}
\arguments{
\item{x}{fitted nproc object using \code{nproc}.}

\item{...}{additional arguments.}
}
\description{
Plot the nproc band(s).
}
\examples{
n = 1000
x = matrix(rnorm(n*2),n,2)
c = 1+3*x[,1]
y = rbinom(n,1,1/(1+exp(-c)))
fit = nproc(x, y, method = 'lda')
plot(fit)
}
\seealso{
\code{\link{npc}}, \code{\link{nproc}}
}

