% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.npc.R
\name{print.npc}
\alias{print.npc}
\title{Print the npc object.}
\usage{
\method{print}{npc}(x, ...)
}
\arguments{
\item{x}{fitted npc object using \code{npc}.}

\item{...}{additional arguments.}
}
\description{
Print the npc object.
}
\examples{
n = 1000
x = matrix(rnorm(n*2),n,2)
c = 1+3*x[,1]
y = rbinom(n,1,1/(1+exp(-c)))
fit = npc(x, y, method = 'lda')
print(fit)
}
\seealso{
\code{\link{npc}}, \code{\link{nproc}}
}

