% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nproc.R
\name{print.nproc}
\alias{print.nproc}
\title{Print the nproc object.}
\usage{
\method{print}{nproc}(x, ...)
}
\arguments{
\item{x}{fitted nproc object using \code{nproc}.}

\item{...}{additional arguments.}
}
\description{
Print the nproc object.
}
\examples{
n = 1000
x = matrix(rnorm(n*2),n,2)
c = 1+3*x[,1]
y = rbinom(n,1,1/(1+exp(-c)))
fit = nproc(x, y, method = 'lda')
print(fit)
}
\seealso{
\code{\link{npc}}, \code{\link{nproc}}
}
