% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.nproc.R
\name{lines.nproc}
\alias{lines.nproc}
\title{Add NP-ROC curves to the current plot object.}
\usage{
\method{lines}{nproc}(x, ...)
}
\arguments{
\item{x}{fitted NP-ROC object using \code{nproc}.}

\item{...}{additional arguments.}
}
\description{
Add NP-ROC curves to the current plot object.
}
\seealso{
\code{\link{npc}}, \code{\link{nproc}} and \code{\link{plot.nproc}}.
}

