% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.nproc.R
\name{plot.nproc}
\alias{plot.nproc}
\title{Plot the nproc curve(s).}
\usage{
\method{plot}{nproc}(x, ...)
}
\arguments{
\item{...}{additional arguments.}

\item{object}{fitted nproc object using \code{nproc}.}
}
\description{
Plot the nproc curve(s).
}
\examples{
n = 1000
x = matrix(rnorm(n*2),n,2)
c = 1+3*x[,1]
y = rbinom(n,1,1/(1+exp(-c)))
fit = nproc(x, y, methods = 'svm')
plot(fit)
fit = nproc(x, y, methods = c('svm','logistic'))
plot(fit)
##Compare Confidence Curves
fit = nproc(x, y, methods = c('svm','logistic','lda'), conf = TRUE)
plot(fit)
}
\seealso{
\code{\link{npc}} and \code{\link{nproc}}
}

